backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
slack-sdk==3.23.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msgraph-sdk==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
textblob==0.17.1
nltk==3.8.1
prometheus-client==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
httpx==0.25.1